- code quality.

+++ tiny-db/src/main/java/org/bambrikii/tiny/db/storage/types/IntIOUtils.java
@@ -28,6 +28,6 @@ public class IntIOUtils {
         channel.read(buff);
         var available = buff.remaining();
         buff.flip();
-        return buff.capacity() - available - 1;
+        return (long) buff.capacity() - available - 1;
     }
 }
